refactor: (#685) PostListPage를 삭제하고 HomePage와 병합 및 적용
