+++ frontend/src/hooks/context/postOption.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, PropsWithChildren, SetStateAction, createContext, useState } from 'react';
 
-import { PostSorting, PostStatus } from '@pages/Home/PostListPage/types';
+import { PostSorting, PostStatus } from '@pages/HomePage/types';
 
 import { SORTING, STATUS } from '@constants/post';
