--- frontend/src/pages/Home/index.tsx
@@ -1,13 +1,0 @@
-import AppInstallPrompt from '@components/common/AppInstallPrompt';
-import Layout from '@components/common/Layout';
-
-import PostListPage from './PostListPage';
-
-export default function Home() {
-  return (
-    <Layout isSidebarVisible={true} isMobileDefaultHeaderVisible={false}>
-      <PostListPage />
-      <AppInstallPrompt />
-    </Layout>
-  );
-}
