+++ frontend/src/pages/HomePage/HomePage.stories.tsx
@@ -1,0 +1,14 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import HomePage from '.';
+
+const meta: Meta<typeof HomePage> = {
+  component: HomePage,
+};
+
+export default meta;
+type Story = StoryObj<typeof HomePage>;
+
+export const Default: Story = {
+  render: () => <HomePage />,
+};
