+++ frontend/src/pages/HomePage/constants.ts
