+++ frontend/src/pages/HomePage/style.ts
