+++ frontend/src/pages/HomePage/types.ts
